:recycle: Convert useTravelAuthorizationSummary to TypeScript.
